Issue 78: Remove "GetUser" interface Rename "GetUserFromDatabase" query as "GetUser"
